chore: Mise à jour des liste dynamiquement
